Move client code to NS API
