+++ Cargo.toml
@@ -61,7 +61,6 @@ members = [
     "common/ip-packet-requests",
     "common/ledger",
     "common/mixnode-common",
-    "common/models",
     "common/network-defaults",
     "common/node-tester-utils",
     "common/nonexhaustive-delayqueue",
@@ -155,7 +154,6 @@ members = [
 default-members = [
     "clients/native",
     "clients/socks5",
-    "common/models",
     "explorer-api",
     "gateway",
     "mixnode",
