--- common/models/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "nym-common-models"
-version = "0.1.0"
-authors.workspace = true
-repository.workspace = true
-homepage.workspace = true
-documentation.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-readme.workspace = true
-
-[dependencies]
-anyhow = { workspace = true }
-bincode = { workspace = true }
-nym-crypto = { path = "../crypto", features = ["asymmetric", "serde"] }
-serde = { workspace = true, features = ["derive"] }
