+++ nym-node-status-api/Cargo.toml
@@ -23,7 +23,7 @@ envy = { workspace = true }
 futures-util = { workspace = true }
 moka = { workspace = true, features = ["future"] }
 nym-bin-common = { path = "../common/bin-common", features = ["models"]}
-nym-common-models = { path = "../common/models" }
+nym-http-api-client = { path = "../common/http-api-client" }
 nym-crypto = { path = "../common/crypto", features = ["asymmetric", "serde"] }
 nym-explorer-client = { path = "../explorer-api/explorer-client" }
 nym-network-defaults = { path = "../common/network-defaults" }
@@ -62,3 +62,11 @@ sqlx = { workspace = true, features = [
   "macros",
   "migrate",
 ] }
+
+[lib]
+name = "node_status_api_client"
+path = "src/client/lib.rs"
+
+[[bin]]
+name = "node_status_api_server"
+path = "src/main.rs"
