+++ nym-node-status-api/src/client/models.rs
@@ -1,7 +1,9 @@
-use nym_crypto::asymmetric::ed25519::{PublicKey, Signature, SignatureError};
+use nym_crypto::asymmetric::ed25519::{PublicKey, Signature};
 use serde::{Deserialize, Serialize};
 
 pub mod get_testrun {
+    use crate::auth::SignedRequest;
+
     use super::*;
     #[derive(Debug, Clone, Deserialize, Serialize)]
     pub struct Payload {
@@ -40,6 +42,8 @@ pub struct TestrunAssignment {
 }
 
 pub mod submit_results {
+    use crate::auth::SignedRequest;
+
     use super::*;
     #[derive(Debug, Clone, Deserialize, Serialize)]
     pub struct Payload {
@@ -69,33 +73,4 @@ pub mod submit_results {
             &self.payload
         }
     }
-}
-
-pub trait SignedRequest {
-    type Payload: serde::Serialize;
-
-    fn public_key(&self) -> &PublicKey;
-    fn signature(&self) -> &Signature;
-    fn payload(&self) -> &Self::Payload;
-}
-
-pub trait VerifiableRequest: SignedRequest {
-    type Error: From<bincode::Error> + From<SignatureError>;
-
-    fn verify_signature(&self) -> Result<(), Self::Error> {
-        bincode::serialize(self.payload())
-            .map_err(Self::Error::from)
-            .and_then(|serialized| {
-                self.public_key()
-                    .verify(serialized, self.signature())
-                    .map_err(Self::Error::from)
-            })
-    }
-}
-
-impl<T> VerifiableRequest for T
-where
-    T: SignedRequest,
-{
-    type Error = anyhow::Error;
 }
